report_service/requirements.txt
fastapi
uvicorn
pymysql
psycopg2-binary